Added the database, hashmap, and other edits
Finished up making a function to get csv file data into hashmap, and another helper function to check user input's symbol is valid and is in the hashmap. Reorganized the code to make it easier to read.  It should work fine now.

There could be edits with the 2-3 timestamp helper functions looks a bit redundant. 

Will need to make a try catch block  for the case where some stocks' IPO date was less than 2 years ago within the main class. Aside from that, the MVP for the back end seems done.